Validaciones

+++ Backend/src/main/java/com/app/logistica/dtos/Package/PackageRequest.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 @Builder
 
 public class PackageRequest {
-    private float weight;
-    private float length;
-    private float width;
-    private float height;
+    private double weight;
+    private double length;
+    private double width;
+    private double height;
 }
